fix useTradePageVersion
